Realtime-Voice/websocket-minimal/python/requirements.txt
websocket-client>=1.6.0
python-dotenv>=1.0.0
